+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/
 ## 0.48.0
 
 - Basic support for collection queryables: just pass through first non-error upstream backend response ([#198](https://github.com/Open-EO/openeo-aggregator/issues/198))
+- Pertitioned job tracking: move sub-job to error state on `describe_job` failure to avoid partitioned job being stuck in "running" ([#200](https://github.com/Open-EO/openeo-aggregator/issues/200))
 
 
 ## 0.47.0